Split index-files lib
